Adapt CCS & DRY Methodology
